third_party/blink/web_tests/platform/generic/virtual/css-calc-infinity-and-nan-disabled/external/wpt/css/css-values/calc-infinity-nan-computed-expected.txt

This is a testharness.js-based test.
FAIL Property width value 'calc(NaN * 1px)' assert_true: 'calc(NaN * 1px)' is a supported value for width. expected true got false
FAIL Property width value 'calc(infinity * 1px)' assert_true: 'calc(infinity * 1px)' is a supported value for width. expected true got false
FAIL Property width value 'calc(infinity * 1cm)' assert_true: 'calc(infinity * 1cm)' is a supported value for width. expected true got false
FAIL Property width value 'calc(NaN * 1rem)' assert_true: 'calc(NaN * 1rem)' is a supported value for width. expected true got false
FAIL Property width value 'calc(10.135262721212548pc - 199pt / NaN)' assert_true: 'calc(10.135262721212548pc - 199pt / NaN)' is a supported value for width. expected true got false
FAIL Property width value 'max(15px, NaN * 1px)' assert_true: 'max(15px, NaN * 1px)' is a supported value for width. expected true got false
FAIL Property width value 'max(NaN * 1px, 15px)' assert_true: 'max(NaN * 1px, 15px)' is a supported value for width. expected true got false
FAIL Property width value 'min(15px, NaN * 1px)' assert_true: 'min(15px, NaN * 1px)' is a supported value for width. expected true got false
FAIL Property width value 'min(NaN * 1px, 15px)' assert_true: 'min(NaN * 1px, 15px)' is a supported value for width. expected true got false
FAIL Property width value 'calc(infinity * 1px - infinity * 1%)' assert_true: 'calc(infinity * 1px - infinity * 1%)' is a supported value for width. expected true got false
FAIL Property width value 'calc(infinity * 1px + infinity * 1%)' assert_true: 'calc(infinity * 1px + infinity * 1%)' is a supported value for width. expected true got false
FAIL Property width value 'calc(min(NaN * 1px, infinity * 1px) + max(infinity * 1px, -infinity * 1px))' assert_true: 'calc(min(NaN * 1px, infinity * 1px) + max(infinity * 1px, -infinity * 1px))' is a supported value for width. expected true got false
FAIL Property width value 'calc(infinity * 1px - max(infinity * 1%, 0%))' assert_true: 'calc(infinity * 1px - max(infinity * 1%, 0%))' is a supported value for width. expected true got false
FAIL Property width value 'calc(max(infinity * 1px, 10px))' assert_true: 'calc(max(infinity * 1px, 10px))' is a supported value for width. expected true got false
FAIL Property margin-left value 'calc(-infinity * 1px)' assert_true: 'calc(-infinity * 1px)' is a supported value for margin-left. expected true got false
FAIL Property margin-left value 'calc(min(1px, -infinity * 1%))' assert_true: 'calc(min(1px, -infinity * 1%))' is a supported value for margin-left. expected true got false
FAIL Property margin-left value 'calc(-infinity * 1%)' assert_true: 'calc(-infinity * 1%)' is a supported value for margin-left. expected true got false
FAIL Property margin-left value 'calc(max(10000px, 0px) + min(-infinity * 1px, infinity * 1px))' assert_true: 'calc(max(10000px, 0px) + min(-infinity * 1px, infinity * 1px))' is a supported value for margin-left. expected true got false
FAIL Property margin-left value 'calc(-infinity * 1px - infinity * 1px)' assert_true: 'calc(-infinity * 1px - infinity * 1px)' is a supported value for margin-left. expected true got false
FAIL Property margin-left value 'calc(min(-infinity * 1px, 10px))' assert_true: 'calc(min(-infinity * 1px, 10px))' is a supported value for margin-left. expected true got false
FAIL Property animation-duration value 'calc(NaN * 1s)' assert_true: 'calc(NaN * 1s)' is a supported value for animation-duration. expected true got false
FAIL Property animation-duration value 'calc(infinity * 1s)' assert_true: 'calc(infinity * 1s)' is a supported value for animation-duration. expected true got false
FAIL Property animation-duration value 'calc(1 / 0 * 1s)' assert_true: 'calc(1 / 0 * 1s)' is a supported value for animation-duration. expected true got false
FAIL Property animation-duration value 'calc(max(infinity * 1s, 10s)' assert_true: 'calc(max(infinity * 1s, 10s)' is a supported value for animation-duration. expected true got false
FAIL Property transition-delay value 'calc(-infinity* 1s)' assert_true: 'calc(-infinity* 1s)' is a supported value for transition-delay. expected true got false
FAIL Property transition-delay value 'calc(max(10000s, 0s) + min(-infinity * 1s, infinity * 1s))' assert_true: 'calc(max(10000s, 0s) + min(-infinity * 1s, infinity * 1s))' is a supported value for transition-delay. expected true got false
FAIL Property transition-delay value 'calc(min(-infinity * 1s, 10s))' assert_true: 'calc(min(-infinity * 1s, 10s))' is a supported value for transition-delay. expected true got false
FAIL Property rotate(calc(infinity * 1deg)) value expected same with rotate(0deg) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(-infinity * 1deg)) value expected same with rotate(0deg) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(NaN * 1deg)) value expected same with rotate(0deg) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(infinity * 1turn)) value expected same with rotate(0turn) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(-infinity * 1turn)) value expected same with rotate(0turn) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(NaN * 1turn)) value expected same with rotate(0turn) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(infinity * 1rad)) value expected same with rotate(0rad) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(-infinity * 1rad)) value expected same with rotate(0rad) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(NaN * 1rad)) value expected same with rotate(0rad) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(infinity * 1grad)) value expected same with rotate(0grad) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(-infinity * 1grad)) value expected same with rotate(0grad) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
FAIL Property rotate(calc(NaN * 1grad)) value expected same with rotate(0grad) in +/-0.0001 assert_array_approx_equals: property 0, expected 1 +/- 0.0001, expected 1 but got NaN
Harness: the test ran to completion.
